atsamd_hal_macros

Attribute Macro hal_macro_helper

#[hal_macro_helper]

Helper macro to allow using #[hal_cfg(..)] macro in more places

Normally the #[cfg(..)] macro is allowed in many more places than proc-macros, such as directly on a statement. This mitigates that restriction.

It can be used like this:

#[hal_macro_helper]
struct MyStruct {
    #[hal_cfg("sercom0")]
    my_field: String
}

This works, because attributes are allowed on the outer item.

The #[hal_macro_helper] will search through the item and replace all instances of the #[hal_cfg(..)] attribute with the corresponding #[cfg(..)] attribute instead. This way the inner attributes are technically not interpreted as a proc-macro at all.