atsamd_hal_macros

Attribute Macro hal_module

#[hal_module]

Macro which expands to a mod foo; item with different paths for each device.

It can be used like this:

#[hal_module(
    any("nvmctrl-d11", "nvmctrl-d21") => "calibration/d11.rs",
    "nvmctrl-d5x" => "calibration/d5x.rs",
)]
pub mod calibration {}

#[hal_module("aes")]
pub mod aes {}

This will then expand to something of the form:

#[cfg(any(feature = "samd11c", ...))]
#[path = "calibration/d11.rs"]
pub mod calibration;

#[cfg(any(feature = "samd51g", ...))]
#[path = "calibration/d5x.rs"]
pub mod calibration;

#[cfg(any(feature = "samd51g", ...))]
pub mod aes;

Ideally you would be to write pub mod calibration; instead of pub mod calibration {}, but unfortunately non-inline modules are not currently supposed in proc macros. See rust#54727 for details.